requirements.txt
streamlit==1.31.0
groq==0.11.0
python-dotenv==1.0.0
textblob==0.17.1
langdetect==1.0.9
pandas==2.1.4
streamlit-chat==0.1.1
nltk